refactor: extract session submission from AgentWorkspace
